reorganized code.
